requirements.txt
fonttools==4.39.3
imageio==2.27.0
ipykernel==6.22.0
ipython==8.12.0
jupyter_client==8.1.0
jupyter_core==5.3.0
lazy_loader==0.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
networkx==3.0
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.5.0
pyparsing==3.0.9
pytesseract==0.3.10
scikit-image==0.20.0
scipy==1.10.1
tesseract==0.1.3